Move from nose2 to pytest.
This commit changes the default test runner from nose2 to mailman.client. It
also adds the functionality that was previously implemented to run tests using
vcrpy and nose2 using pytest plugins.

pytest-vcr can be used to run tests using VCR.py and adds a custom command line
option --vcr-record-mode which can be used to change the recording mode for the
VCR.py.