Always quote special characters in emails when used in URL.
Emails can contains special characters, which when used in URLs, can cause
errors because of Core interpreting them in a way which we don't intend to.

For example a `?` in an email address, which is a valid character, returns a 404
error since everything after a literal `?` is expected to be query parameters.