Add Connection hooks in Mailmanclient to access final request params
This adds a generic mechanism to hook into the behavior of Mailmanclient by
letting hooks manipulate the request parameters. The primary use case of this
feature as of now is to simply log all connections to Mailman's REST API so
that we can determine how many calls does each view make to Core.

Fixes #68